+++ multiboot2/src/boot_loader_name.rs
@@ -3,7 +3,6 @@
 use crate::tag::TagHeader;
 use crate::{StringError, TagType, parse_slice_as_string};
 use core::fmt::{Debug, Formatter};
-use core::mem;
 use multiboot2_common::{MaybeDynSized, Tag};
 #[cfg(feature = "builder")]
 use {alloc::boxed::Box, multiboot2_common::new_boxed};
@@ -78,7 +77,7 @@ impl Debug for BootLoaderNameTag {
 impl MaybeDynSized for BootLoaderNameTag {
     type Header = TagHeader;
 
-    const BASE_SIZE: usize = mem::size_of::<TagHeader>();
+    const BASE_SIZE: usize = size_of::<TagHeader>();
 
     fn dst_len(header: &TagHeader) -> usize {
         assert!(header.size as usize >= Self::BASE_SIZE);